Update README.md
Create a class in JobApplicant.java that holds data about a job applicant. Include a name, a phone, and four Boolean fields, hasWordSkill, hasSpreadsheetSkill, hasDatabaseSkill, and hasGraphicsSkill, that represent whether the applicant is skilled in each of the following areas: word processing, spreadsheets, databases, and graphics. Include a constructor that accepts values for each of the fields. Also implement a get method for each field. The get method should be the field name prefixed with 'get'. For example, the get method for name should be called getName.

Create an application in TestJobApplicants.java that instantiates several job applicant objects and passes each in turn to a Boolean method named isQualified that determines whether each applicant is qualified for an interview. The method should return true if the applicant has at least three of the four skills, false otherwise. Then, in the main() method, display an appropriate message for each applicant.